dj - rearranged a bunch of stuff, three total levels of inheritance feels like a bit much
