+++ src/test/java/edu/ucsb/cs156/example/web/HomePageWebIT.java
@@ -2,46 +2,10 @@ package edu.ucsb.cs156.example.web;
 
 import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
 
-import com.microsoft.playwright.Browser;
-import com.microsoft.playwright.BrowserContext;
-import com.microsoft.playwright.BrowserType;
-import com.microsoft.playwright.Page;
-import com.microsoft.playwright.Playwright;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
+import edu.ucsb.cs156.example.WebTestCase;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.context.ActiveProfiles;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("integration")
-public class HomePageWebIT {
-  @Value("${app.playwright.headless:true}")
-  private boolean runHeadless;
-
-  @LocalServerPort private int port;
-
-  private Browser browser;
-  private Page page;
-
-  @BeforeEach
-  public void setup() {
-    browser =
-        Playwright.create()
-            .chromium()
-            .launch(new BrowserType.LaunchOptions().setHeadless(runHeadless));
-
-    BrowserContext context = browser.newContext();
-    page = context.newPage();
-  }
-
-  @AfterEach
-  public void teardown() {
-    browser.close();
-  }
-
+public class HomePageWebIT extends WebTestCase {
   @Test
   public void home_page_shows_greeting() throws Exception {
     String url = String.format("http://localhost:%d/", port);
